+++ src/cli.ts
@@ -17,6 +17,14 @@ const main = defineCommand({
       description: "The directory to build",
       required: false,
     },
+    config: {
+      type: "string",
+      description: [
+        "The configuration file to use relative to the current working directory.",
+        "                 Unbuild tries to read `build.config` from the build `DIR` by default.",
+        "",
+      ].join("\n"),
+    },
     watch: {
       type: "boolean",
       description: "Watch the src dir and rebuild on change (experimental)",
@@ -38,6 +46,7 @@ const main = defineCommand({
     const rootDir = resolve(process.cwd(), args.dir || ".");
     await build(rootDir, args.stub, {
       sourcemap: args.sourcemap,
+      config: args.config ? resolve(args.config) : undefined,
       stub: args.stub,
       watch: args.watch,
       rollup: {