+++ test/validate.test.ts
@@ -1,9 +1,9 @@
+import type { BuildEntry } from "../src/types.ts";
 import { fileURLToPath } from "node:url";
 import { consola } from "consola";
 import { join } from "pathe";
 import { describe, it, expect } from "vitest";
 import { validateDependencies, validatePackage } from "../src/validate";
-import { BuildEntry } from "../src/types";
 
 describe("validatePackage", () => {
   it("detects missing files", () => {
@@ -60,6 +60,7 @@ describe("validateDependencies", () => {
         stub: false,
         alias: {},
         replace: {},
+        // @ts-expect-error
         rollup: {
           replace: false,
           alias: false,
@@ -79,7 +80,11 @@ describe("validateDependencies", () => {
   it("does not print implicit deps warning for peerDependencies", () => {
     const logs: string[] = [];
     consola.mockTypes((type) =>
-      type === "warn" ? (str: string) => logs.push(str) : () => {},
+      type === "warn"
+        ? (str: string): void => {
+            logs.push(str);
+          }
+        : (): void => {},
     );
 
     validateDependencies({
@@ -99,6 +104,7 @@ describe("validateDependencies", () => {
         stub: false,
         alias: {},
         replace: {},
+        // @ts-expect-error
         rollup: {
           replace: false,
           alias: false,